+++ packages/fluentui/react/src/themes/teams/components/Embed/embedStyles.ts
@@ -3,6 +3,8 @@ import { pxToRem } from '../../../../utils';
 import Embed, { EmbedProps, EmbedState } from '../../../../components/Embed/Embed';
 import { EmbedVariables } from './embedVariables';
 import getBorderFocusStyles from '../../getBorderFocusStyles';
+import playIndicatorUrl from './playIndicatorUrl';
+import pauseIndicatorUrl from './pauseIndicatorUrl';
 
 export default {
   root: ({ props: p, variables: v, theme: { siteVariables } }): ICSSInJSStyle => {
@@ -37,7 +39,20 @@ export default {
   control: ({ props: p, variables: v }): ICSSInJSStyle => ({
     background: `0 no-repeat ${v.controlBackgroundColor}`,
     backgroundPositionX: pxToRem(3),
+    width: pxToRem(48),
+    height: pxToRem(48),
     color: v.controlColor,
+
+    padding: pxToRem(4),
+    borderRadius: '50%',
+
+    backgroundImage: playIndicatorUrl(v.controlColor),
+    backgroundRepeat: 'no-repeat',
+    backgroundPosition: 'center',
+
+    ...(p.active && {
+      backgroundImage: pauseIndicatorUrl(v.controlColor)
+    }),
 
     opacity: p.active ? 0 : 1,
     pointerEvents: 'none',